+++ build.clj
@@ -23,10 +23,6 @@
 (defn jar [_]
   (clean nil)
   (println "Building" lib version)
-  (b/compile-clj {:basis      jar-basis
-                  :class-dir  class-dir
-                  :src-dir    ["src"]
-                  :ns-compile ['com.eldrix.hades.core]})
   (b/write-pom {:class-dir class-dir
                 :lib       lib
                 :version   version